moved lreshape and wide_to_long
